--user -unset-group unshare projects owned by other members

+++ lib/inotify/SharingStructureMXF.php
@@ -98,6 +98,10 @@ class SharingStructureMXF
 		foreach($folders as $folder)
 		{
 			$linkName="$mxfFolder/$folder";
+
+			//file could have been deleted already by a concurrent process
+			if(!file_exists($linkName)) continue;
+
 			$target=readlink($linkName);
 
 
@@ -124,7 +128,8 @@ class SharingStructureMXF
 
 			if(!SharingFolders::isGroupMemberHomeFolder($users,$targetHomeFolder))
 			{
-				unlink($linkName);
+				//file could have been deleted already by a concurrent process
+				if(file_exists($linkName)) unlink($linkName);
 				syslog_notice("Removed '$linkName'; in target '$target' ".
 					"the home folder '$targetHomeFolder' is not the home folder for a group member");
 				return;
